Move configure_logging to utils
- Add configure_logging to utils/logging.py

- Update agents to use configure_logging utility
